# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WebEncode API
  description: |
    WebEncode is a distributed video encoding platform that provides VOD transcoding,
    live streaming management, and multi-platform publishing capabilities.
    
    ## Authentication
    All API endpoints require authentication via Bearer token in the Authorization header.
    Use the Auth Plugin to obtain tokens.
    
    ## Rate Limiting
    API requests are rate limited to 100 requests per minute per IP address.
    Exceeding this limit will result in a 429 Too Many Requests response.
  version: 1.0.0
  contact:
    name: WebEncode Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8090
    description: Local development
  - url: https://api.webencode.example.com
    description: Production
...
paths:
...
  /v1/profiles:
    get:
      tags: [Profiles]
      summary: List encoding profiles
      responses:
        '200':
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EncodingProfile'
    post:
      tags: [Profiles]
      summary: Create encoding profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Profile created
  /v1/profiles/{id}:
    get:
      tags: [Profiles]
      summary: Get profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profile details
        '404':
          description: Profile not found
    delete:
      tags: [Profiles]
      summary: Delete profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Profile deleted
...
  /v1/plugins/{id}/enable:
    post:
      tags: [Plugins]
      summary: Enable a plugin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plugin enabled
  /v1/plugins/{id}/disable:
    post:
      tags: [Plugins]
      summary: Disable a plugin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plugin disabled
...
components:
...
  schemas:
...
    EncodingProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        video_codec:
          type: string
        audio_codec:
          type: string
        width:
          type: integer
        height:
          type: integer
        bitrate_kbps:
          type: integer
        preset:
          type: string
        container:
          type: string
    CreateProfileRequest:
      type: object
      required:
        - name
        - video_codec
      properties:
        name:
          type: string
        video_codec:
          type: string
        audio_codec:
          type: string
        width:
          type: integer
        height:
          type: integer
        bitrate_kbps:
          type: integer
        preset:
          type: string
        container:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []